+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Services/Taxonomies/AbstractTaxonomy.php
@@ -70,27 +70,28 @@ abstract class AbstractTaxonomy extends AbstractAutomaticallyInstantiatedService
             'show_in_nav_menus' => true,
             'show_tagcloud' => false,
             'show_in_rest' => true,
+            'show_admin_column' => $this->showAdminColumn(),
         );
         /**
-         * From documentation concerning 2nd parameter ($object_type)
-         * > Setting explicitly to null registers the taxonomy but doesn't associate it
-         * > with any objects, so it won't be directly available within the Admin UI.>
-         * > You will need to manually register it using the 'taxonomy' parameter
-         * > (passed through $args)
-         * > when registering a custom post_type
+         * From documentation concerning 2nd parameter ($object_type):
+         *
+         *   > Setting explicitly to null registers the taxonomy but doesn't associate it
+         *   > with any objects, so it won't be directly available within the Admin UI.>
+         *   > You will need to manually register it using the 'taxonomy' parameter
+         *   > (passed through $args)
+         *   > when registering a custom post_type
+         *
          * @see https://codex.wordpress.org/Function_Reference/register_taxonomy#Parameters
          */
         \register_taxonomy(
             $this->getTaxonomy(),
-            [],
+            $this->getCustomPostTypes(),
             $args
         );
     }
 
-    /**
-     * Taxonomy plural name
-     *
-     * @param bool $titleCase Indicate if the name must be title case (for starting a sentence) or, otherwise, lowercase
-     */
-    abstract protected function getTaxonomyPluralNames(bool $titleCase = true): string;
+    protected function showAdminColumn(): bool
+    {
+        return $this->isHierarchical();
+    }
 }