+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Services/Taxonomies/TaxonomyInterface.php
@@ -10,5 +10,17 @@ interface TaxonomyInterface
 
     public function getTaxonomyName(bool $titleCase = true): string;
 
+    /**
+     * Taxonomy plural name
+     *
+     * @param bool $titleCase Indicate if the name must be title case (for starting a sentence) or, otherwise, lowercase
+     */
+    public function getTaxonomyPluralNames(bool $titleCase = true): string;
+
+    /**
+     * @return string[]
+     */
+    public function getCustomPostTypes(): array;
+
     public function isHierarchical(): bool;
 }